apps/api/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.1,<4.0
requests>=2.32,<3.0
gunicorn>=21.2,<22.0
drf-spectacular[sidecar]>=0.27,<0.28
cryptography>=43.0,<44.0
PyJWT>=2.9,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=4.9,<6.0
boto3>=1.34,<2.0
